refactor ConsentScreen
