+++ src/features/register/ConsentScreen/index.tsx
@@ -1,0 +1,2 @@
+import ConsentScreen from './ConsentScreen';
+export { ConsentScreen };
